+++ tests/integration/test_states.py
@@ -2,15 +2,15 @@
 Integration tests for states.
 '''
 
-import unittest as ut
 import numpy as np
 
 import dynamite_test_runner as dtr
 
 from dynamite import config
-from dynamite.states import State
+from dynamite.states import State, UninitializedError
 from dynamite.subspaces import Parity, SpinConserve, Auto
 from dynamite.operators import sigmaz, sigmax, sigmay, index_sum
+from dynamite.computations import reduced_density_matrix
 
 class RandomSeed(dtr.DynamiteTestCase):
 
@@ -69,6 +69,7 @@ class PetscMethods(dtr.DynamiteTestCase):
         state.vec[start:end] = np.array([1]*(end-start))
         state.vec.assemblyBegin()
         state.vec.assemblyEnd()
+        state.set_initialized()
         self.assertAlmostEqual(state.norm()**2, state.subspace.get_dimension())
 
     def test_normalize(self):
@@ -77,6 +78,8 @@ class PetscMethods(dtr.DynamiteTestCase):
         state.vec[start:end] = np.array([1]*(end-start))
         state.vec.assemblyBegin()
         state.vec.assemblyEnd()
+        state.set_initialized()
+
         state.normalize()
         self.assertTrue(state.norm() == 1)
 
@@ -87,6 +90,7 @@ class PetscMethods(dtr.DynamiteTestCase):
         state1.vec[start:end] = np.arange(start, end)
         state1.vec.assemblyBegin()
         state1.vec.assemblyEnd()
+        state1.set_initialized()
 
         result = np.ndarray((end-start,), dtype=np.complex128)
         state1.copy(state2)
@@ -107,6 +111,7 @@ class PetscMethods(dtr.DynamiteTestCase):
         state1.vec[start:end] = np.arange(start, end)
         state1.vec.assemblyBegin()
         state1.vec.assemblyEnd()
+        state1.set_initialized()
 
         result = np.ndarray((end-start,), dtype=np.complex128)
         state2 = state1.copy()
@@ -142,13 +147,13 @@ class PetscMethods(dtr.DynamiteTestCase):
 
     def test_scale_exception_ary(self):
         val = np.array([3.1, 4])
-        state = State()
+        state = State(state='U'*config.L)
         with self.assertRaises(TypeError):
             state *= val
 
     def test_scale_exception_vec(self):
-        state1 = State()
-        state2 = State()
+        state1 = State(state='U'*config.L)
+        state2 = State(state='U'*config.L)
         with self.assertRaises(TypeError):
             state1 *= state2
 
@@ -244,6 +249,83 @@ class Saving(dtr.DynamiteTestCase):
         state.save(self.fname)
         loaded = State.from_file(self.fname)
         self.check_states_equal(state, loaded)
+
+
+class Uninitialized(dtr.DynamiteTestCase):
+
+    def test_copy_neither(self):
+        s0 = State()
+        s1 = s0.copy()
+        self.assertFalse(s1.initialized)
+
+    def test_copy_from_uninit(self):
+        s0 = State()
+        s1 = State(state='U'*config.L)
+        with self.assertRaises(UninitializedError):
+            s0.copy(s1)
+
+    def test_copy_to_uninit(self):
+        s0 = State(state='U'*config.L)
+        s1 = State()
+        s0.copy(s1)
+        self.assertTrue(s1.initialized)
+
+    def test_project(self):
+        s0 = State()
+        with self.assertRaises(UninitializedError):
+            s0.project(0, 0)
+
+    def test_to_numpy(self):
+        s0 = State()
+        with self.assertRaises(UninitializedError):
+            s0.to_numpy()
+
+    def test_dot_from(self):
+        s0 = State()
+        s1 = State(state='U'*config.L)
+        with self.assertRaises(UninitializedError):
+            s0.dot(s1)
+
+    def test_dot_to(self):
+        s0 = State(state='U'*config.L)
+        s1 = State()
+        with self.assertRaises(UninitializedError):
+            s0.dot(s1)
+
+    def test_norm(self):
+        s0 = State()
+        with self.assertRaises(UninitializedError):
+            s0.norm()
+
+    def test_normalize(self):
+        s0 = State()
+        with self.assertRaises(UninitializedError):
+            s0.normalize()
+
+    def test_imul(self):
+        s0 = State()
+        with self.assertRaises(UninitializedError):
+            s0 *= 2
+
+    def test_idiv(self):
+        s0 = State()
+        with self.assertRaises(UninitializedError):
+            s0 /= 2
+
+    def test_op_dot(self):
+        s0 = State()
+        with self.assertRaises(UninitializedError):
+            sigmax().dot(s0)
+
+    def test_op_evolve(self):
+        s0 = State()
+        with self.assertRaises(UninitializedError):
+            sigmax().evolve(s0, t=1.0)
+
+    def test_rdm(self):
+        s0 = State()
+        with self.assertRaises(UninitializedError):
+            reduced_density_matrix(s0, [])
 
 # TODO: check state setting. e.g. setting an invalid state should fail (doesn't for Full subspace)