+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - `Subspace.identical` method determines whether two subspaces are exactly the same (both equal and of the same type)
  - `Operator.has_subspace` method returns whether a given subspace has been registered with the operator
  - `State.save` and `State.from_file` methods to allow saving states to disk and loading them later
+ - `State.initialized` flag to track whether the state's vector has been filled with data
 
 ### Changed
  - `L` of operators and states is now stored in their subspaces